:fire: CONTRAST-31008 Removing steps which instruct user how to add integration tests
This how-to showed users how to add integration tests to a project using
the maven-cargo-plugin and subsequently how to configure the
maven-cargo-plugin to include Contrast. This change re-focuses the
article on just adding Contrast to an existing maven-cargo-plugin
configuration. We want to move the "how to add integration tests that
work with Contrast" documentation to an explanation article that is goes
into the details of automated testing + Contrast, and we want to keep
this how-to focused on Contrast + maven-cargo-plugin.

Other changes:

* Updated artifactId java-agent --> contrast-agent
* Removed the jar renaming exercise that will be obviated by the changes
  we're making before releasing to Maven Central
* Removed the unit test to focus the user on the integration test and
  remove any ambiguity about which test works with Contrast
* Added comments